refactor: relocate server/infras into routers and usecases
Move better-auth (auth config + drizzle schemas) from server/infras to
server/routers since it is part of the HTTP surface and unrelated to the
usecases/adaptors implementations. Move the remaining infras modules
(ai-sdk, local-hermeum-config, kubernetes) under server/usecases/infras
to sit next to the adaptor interfaces they implement.

Update all import paths, vi.mock specifiers, drizzle configs, and the
better-auth:generate script accordingly.
