+++ apps/dashboard/drizzle.config.postgres.ts
@@ -2,7 +2,7 @@ import { defineConfig } from "drizzle-kit";
 
 export default defineConfig({
   dialect: "postgresql",
-  schema: ["./src/server/infras/better-auth/schema.postgres.ts"],
+  schema: ["./src/server/routers/better-auth/schema.postgres.ts"],
   dbCredentials: {
     url: process.env.HERMEUM_DATABASE_URL!,
   },
