+++ apps/dashboard/drizzle.config.sqlite.ts
@@ -2,7 +2,7 @@ import { defineConfig } from "drizzle-kit";
 
 export default defineConfig({
   dialect: "sqlite",
-  schema: ["./src/server/infras/better-auth/schema.sqlite.ts"],
+  schema: ["./src/server/routers/better-auth/schema.sqlite.ts"],
   dbCredentials: {
     url: process.env.HERMEUM_DATABASE_URL!,
   },
