+++ apps/dashboard/package.json
@@ -9,7 +9,7 @@
     "start": "NODE_ENV=production tsx src/server/server.ts",
     "lint": "eslint .",
     "typecheck": "tsc --noEmit",
-    "better-auth:generate": "pnpm dlx auth generate --config src/server/infras/better-auth/auth.ts --output src/server/infras/better-auth/schema.${HERMEUM_DATABASE_DIALECT:-sqlite}.ts --yes",
+    "better-auth:generate": "pnpm dlx auth generate --config src/server/routers/better-auth/auth.ts --output src/server/routers/better-auth/schema.${HERMEUM_DATABASE_DIALECT:-sqlite}.ts --yes",
     "drizzle:generate": "drizzle-kit generate --config drizzle.config.${HERMEUM_DATABASE_DIALECT:-sqlite}.ts",
     "drizzle:migrate": "drizzle-kit migrate --config drizzle.config.${HERMEUM_DATABASE_DIALECT:-sqlite}.ts",
     "clean": "rm -rf dist *.tsbuildinfo",
