+++ apps/dashboard/src/server/routers/better-auth/auth.ts
