+++ apps/dashboard/src/server/usecases/agent.test.ts
@@ -1,8 +1,8 @@
 import { describe, it, expect, vi } from "vitest";
 
-vi.mock("../infras/kubernetes/client", () => ({ KubernetesClient: vi.fn() }));
-vi.mock("../infras/local-hermeum-config", () => ({ LocalConfig: vi.fn() }));
-vi.mock("../infras/ai-sdk", () => ({ AiSdkGenerator: vi.fn() }));
+vi.mock("./infras/kubernetes/client", () => ({ KubernetesClient: vi.fn() }));
+vi.mock("./infras/local-hermeum-config", () => ({ LocalConfig: vi.fn() }));
+vi.mock("./infras/ai-sdk", () => ({ AiSdkGenerator: vi.fn() }));
 
 import { AgentUseCase, AGENT_INPUT_SYSTEM_PROMPT } from "./agent";
 import type { AiGenerator } from "./adaptors/generator";
