+++ apps/dashboard/src/server/usecases/infras/kubernetes/types/hermes-agent.ts
